requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
email-validator==2.0.0
python-dateutil==2.8.2
pytz==2023.3
Werkzeug==2.3.7
gunicorn==21.2.0
